+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## **Unreleased**
+
+- Add support for managing integration services
+
 ## v0.1.1 2026-05-22
 
 - Add support for Ruby 3.0